requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
SQLAlchemy>=1.4,<2.1
python-dotenv>=1.0.1
pytest>=8.0.0
